Fix bug when SessionStore file does not yet exist
and improve code here overall
