+++ app/src/main/java/org/walleth/walletconnect/WalletConnectDriver.kt
@@ -1,16 +1,12 @@
 package org.walleth.walletconnect
 
 import android.content.Context
-import android.content.Intent
-import android.net.Uri
 import com.squareup.moshi.Moshi
 import okhttp3.OkHttpClient
 import okhttp3.Request
 import okhttp3.RequestBody
 import okhttp3.Response
 import org.json.JSONObject
-import org.kethereum.erc681.ERC681
-import org.kethereum.erc681.generateURL
 import org.kethereum.model.Address
 import org.ligi.tracedroid.logging.Log
 import org.spongycastle.crypto.digests.SHA256Digest
@@ -21,109 +17,24 @@ import org.spongycastle.crypto.paddings.PKCS7Padding
 import org.spongycastle.crypto.paddings.PaddedBufferedBlockCipher
 import org.spongycastle.crypto.params.KeyParameter
 import org.spongycastle.crypto.params.ParametersWithIV
-import org.walleth.activities.WalletConnectSendTransactionActivity
 import org.walleth.data.JSON_MEDIA_TYPE
-import org.walleth.khex.clean0xPrefix
 import org.walleth.khex.hexToByteArray
 import org.walleth.khex.toNoPrefixHexString
+import org.walleth.walletconnect.model.Session
+import org.walleth.walletconnect.model.StatefulWalletConnectTransaction
+import org.walleth.walletconnect.model.WrappedWalletConnectTransaction
 import java.io.File
-import java.math.BigInteger
 import java.security.SecureRandom
-
-
-fun String.isWalletConnectJSON() = try {
-    JSONObject(this).let {
-        it.has("domain") && (it.has("sessionId") && it.has("sharedKey") && it.has("dappName"))
-    }
-} catch (e: Exception) {
-    false
-}
-
-class WalletConnectTransaction(val from: String,
-                               val to: String,
-                               val nonce: String,
-                               val gasPrice: String,
-                               val gasLimit: String,
-                               val gas: String,
-                               val value: String,
-                               val data: String)
-
-class StatefulWalletConnectTransaction(val tx: WalletConnectTransaction,
-                                       val session: Session,
-                                       val id: String)
-
-
-class WrappedWalletConnectTransaction(val data: WalletConnectTransaction)
-
-data class Session(val sessionId: String,
-                   val domain: String,
-                   val dappName: String,
-                   val sharedKey: String)
-
-data class SessionList(val sessions: List<Session>)
-
-const val INTENT_KEY_WCTXID = "WalletConnectTransactionId"
-const val INTENT_KEY_WCSESSIONID = "WalletConnectSessionId"
-
-fun Context.createIntentForTransaction(statefulTransaction: StatefulWalletConnectTransaction): Intent {
-    val tx = statefulTransaction.tx
-    val url = ERC681(scheme = "ethereum",
-            address = tx.to,
-            value = BigInteger(tx.value.clean0xPrefix(), 16),
-            gas = BigInteger(tx.gasLimit.clean0xPrefix(), 16)
-    ).generateURL()
-
-    return Intent(this, WalletConnectSendTransactionActivity::class.java).apply {
-        data = Uri.parse(url)
-        putExtra(INTENT_KEY_WCTXID, statefulTransaction.id)
-        putExtra(INTENT_KEY_WCSESSIONID, statefulTransaction.session.sessionId)
-        putExtra("nonce", tx.nonce)
-        putExtra("data", tx.data)
-        putExtra("gasPrice", tx.gasPrice)
-        putExtra("from", tx.from)
-        putExtra("parityFlow", false)
-    }
-}
-
-class SessionStore(val direcory: File) {
-
-    private val sessionListAdapter = Moshi.Builder().build().adapter(SessionList::class.java)
-
-    private val sessionMap by lazy {
-        mutableMapOf<String, Session>().apply {
-            sessionListAdapter.fromJson(direcory.readText())?.sessions?.forEach { session ->
-                put(session.sessionId, session)
-            }
-        }
-    }
-
-    fun put(session: Session) {
-        sessionMap[session.sessionId] = session
-
-        val sessionJSO = sessionListAdapter.toJson(SessionList(sessionMap.map { it.value }))
-
-        direcory.writeText(sessionJSO)
-    }
-
-    fun get(key: String): Session? {
-        val session = sessionMap[key]
-
-        if (session == null) {
-            Log.w("Cannot find session for id $key in SessionStore")
-        }
-
-        return session
-    }
-}
 
 class WalletConnectDriver(
         private val context: Context,
         private val pushServerURL: String,
         private val okHttpClient: OkHttpClient) {
 
-    val sessionStore by lazy { SessionStore(File(context.cacheDir, "walletconnect_sessionstore.json")) }
-    var fcmToken = ""
     var txAction: ((tx: StatefulWalletConnectTransaction) -> Unit)? = null
+
+    private val sessionStore by lazy { SessionStore(File(context.cacheDir, "walletconnect_sessionstore.json")) }
+    private var fcmToken = ""
 
     private var aad = 1
