+++ app/src/main/java/org/walleth/walletconnect/WalletConnectIDs.kt
@@ -1,0 +1,5 @@
+package org.walleth.walletconnect
+
+const val INTENT_KEY_WCTXID = "WalletConnectTransactionId"
+const val INTENT_KEY_WCSESSIONID = "WalletConnectSessionId"
+
