+++ app/src/main/java/org/walleth/walletconnect/model/Session.kt
@@ -1,0 +1,6 @@
+package org.walleth.walletconnect.model
+
+data class Session(val sessionId: String,
+                   val domain: String,
+                   val dappName: String,
+                   val sharedKey: String)
