+++ app/src/main/java/org/walleth/walletconnect/model/SessionList.kt
@@ -1,0 +1,3 @@
+package org.walleth.walletconnect.model
+
+data class SessionList(val sessions: List<Session>)
