+++ app/src/main/java/org/walleth/walletconnect/model/StatefulWalletConnectTransaction.kt
@@ -1,0 +1,5 @@
+package org.walleth.walletconnect.model
+
+data class StatefulWalletConnectTransaction(val tx: WalletConnectTransaction,
+                                            val session: Session,
+                                            val id: String)
