+++ app/src/main/java/org/walleth/walletconnect/model/WrappedWalletConnectTransaction.kt
@@ -1,0 +1,3 @@
+package org.walleth.walletconnect.model
+
+data class WrappedWalletConnectTransaction(val data: WalletConnectTransaction)
